fix(components): fix cannot update table size by container resized
