+++ .eslintrc
@@ -80,7 +80,7 @@
         ]
       }
     ],
-    "react/display-name": 1,
+    "react/display-name": 0,
     "react/prop-types": 0,
     "react/state-in-constructor": 0,
     "react/jsx-props-no-spreading": 0,
