+++ packages/s2-react/playground/index.tsx
@@ -116,7 +116,7 @@ const Playground: FC = () => {
     data,
   };
 
-  return <SheetComponent dataCfg={s2DataConfig} options={s2options} />;
+  return <SheetComponent dataCfg={s2DataConfig} options={s2options} adaptive />;
 };
 
 ReactDOM.render(<Playground />, document.getElementById('root'));
