+++ packages/s2-react/src/components/sheets/base-sheet/index.less
@@ -3,6 +3,10 @@
 .@{preCls} {
   &-container {
     overflow: hidden;
+
+    canvas {
+      display: block;
+    }
   }
 
   &-pagination {
