+++ packages/s2-react/src/components/sheets/base-sheet/index.tsx
@@ -26,7 +26,7 @@ import {
 import { DrillDown } from '@/components/drill-down';
 import { Header } from '@/components/header';
 import { BaseSheetProps } from '@/components/sheets/interface';
-import { useResizeEffect } from '@/components/sheets/hooks';
+import { useResizeEffect } from '@/hooks';
 import { S2Pagination } from '@/components/pagination';
 import { TooltipComponent } from '@/components/tooltip';
 import { TooltipRenderProps } from '@/components/tooltip/interface';
