+++ packages/s2-react/src/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './usePaginationEffect';
+export * from './useResizeEffect';
